Update patch set 1
Patch Set 1:

> > I doubt it unfortunately. ipaddress on python 2 has a really
 > > strange behavior: it only accepts unicode, str is rejected. We
 > can
 > > give it a try, but I suspect it won't work.
 > 
 > I've tested it locally, and some unit tests failed, so we'll need
 > to pin this package in ussuri-uc until we get the centos8 packages.

hmm ipaddress varies in py2/py3:-
[ykarel@yatinkarel rebuild]$ python -c 'import ipaddress;print(ipaddress.ip_address("192.1.2.3"))'
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/usr/lib/python2.7/site-packages/ipaddress.py", line 165, in ip_address
    ' a unicode object?' % address)
ipaddress.AddressValueError: '192.1.2.3' does not appear to be an IPv4 or IPv6 address. Did you pass in a bytes (str in Python 2) instead of a unicode object?
[ykarel@yatinkarel rebuild]$ python3 -c 'import ipaddress;print(ipaddress.ip_address("192.1.2.3"))'
192.1.2.3

so either it needs to be fixed upstream or we need to pin in RDO. @dtantsur thoughts?

Patch-set: 1