Update patch set 1
Patch Set 1: Code-Review-1

ironic-python-agent is pinned to commit 8e246f4482151ebbce4de552a6bad403aa0f37ea until we get a promotion with centos8 and can get rid of centos7 and python2 so we can't merge this yet.

Also, we need to update python-werkzeug for CentOS8, latest build in fedora is 0.14.1 which is too old. I'll see if we can get it updated in rdo as soon as possible.

Patch-set: 1